Remove proxy Dockerfile and refactor mobile and backend code
The proxy Dockerfile has been deleted because it is no longer needed. Mobile screenshots and assets are now ignored in git. Tokens are no longer deleted upon landing in the mobile app. The admin-dashboard connects to an updated localhost IP address. The email service now supports plain and HTML email bodies. A run-proxy command has also been added to the Makefile in the API gateway. AndroidManifest now accommodates app-specific URIs. This commit is part of a broader codebase cleanup and optimization effort.